% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_alphas.R
\name{estimate_alphas}
\alias{estimate_alphas}
\title{Estimates factor model alphas on sorted portfolios.}
\usage{
estimate_alphas(data, ret, n_portfolios = 10, lag = 6, datevar = MCALDT)
}
\arguments{
\item{data}{A data frame with returns of sorted portfolios, e.g. the output of the \code{univariate_sort} function. The data must contain the FF6 factors as downloaded by the \code{get_ff_factors} function.}

\item{ret}{Return variable, e.g. \code{RET_VW} for value-weighted returns, as returned by \code{univariate_sort}.}

\item{n_portfolios}{The number of portfolios in the input data, default is 10.}

\item{lag}{Lag for the Newey West adjustment of the standard errors.}

\item{datevar}{Date variable, usually MCALDT with monthly data, use YYYYQ for quarterly data.}
}
\value{
A tibble with returns and alphas for the sorted portfolios, ready to print with \code{kable}.
}
\description{
Produce a table with portfolio excess returns and alphas for the factor models of Fama & French (1993, 2015, 2018).
}
\examples{
\dontrun{
size_deciles <- estimate_alphas(sfz_mon, MTCAP, MRETRF)
}
}
