% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rolling_betas.R
\name{estimate_rolling_betas}
\alias{estimate_rolling_betas}
\title{Estimate rolling window betas in the cross section of stocks.}
\usage{
estimate_rolling_betas(data, regmodel, window, freq = "monthly")
}
\arguments{
\item{data}{A data frame with daily or monthly returns for each KYPERMNO and at least one factor to estimate OLS betas.}

\item{regmodel}{A regression formula \code{y~x} to estimate factor betas.}

\item{window}{A \code{dbl} value for the estimation window in months in {1,3,6,12,24,...,60}.}

\item{freq}{Either "monthly" (default) or "daily" for the input frequency of the data.}
}
\value{
A tibble with the estimated betas for all input factors on a monthly basis
}
\description{
Estimate rolling window betas in the cross section of stocks.
}
\examples{
\dontrun{
capm_beta_monthly <- estimate_rolling_betas(sfz_mon, "RETRF~MKTRF", 60, "monthly")
}
}
