% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ff_factors.R
\name{get_ff_factors}
\alias{get_ff_factors}
\title{Download the asset pricing factors of Fama & French.}
\usage{
get_ff_factors(freq = "monthly")
}
\arguments{
\item{freq}{Either "monthly" (default), "daily", or "annual".}
}
\value{
A tibble with the six factors MKTRF, SMB, HML, RMW, CMA, and MOM as documented by Fama & French (1993, 2015, 2018)
}
\description{
Downloads the factors of the Fama & French six factor model from the data library of Prof. Kenneth R. French.
}
\examples{
\dontrun{
ff_factors_monthly <- get_ff_factors(freq = "monthly")
ff_factors_annual <- get_ff_factors(freq = "annual")
ff_factors_daily <- get_ff_factors(freq = "daily")
}
}
\references{
Nelson Areal: \code{frenchdata}; Fama & French (1993, 2015, 2018)
}
