% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_q_factors.R
\name{get_q_factors}
\alias{get_q_factors}
\title{Download the asset pricing factors of Hou et al. (2019)}
\usage{
get_q_factors(freq = "monthly")
}
\arguments{
\item{freq}{Either "monthly" (default), "daily", or "annual".}
}
\value{
A tibble with the five factors MKTRF, ME, IA, ROE, EG as documented by Hou et al. (2019)
}
\description{
Downloads the factors of the Hou et al. (2019) q5 model from the data library q-global.org
}
\examples{
\dontrun{
q_factors_monthly <- get_q_factors(freq = "monthly")
q_factors_annual <- get_q_factors(freq = "annual")
q_factors_daily <- get_q_factors(freq = "daily")
}
}
\references{
Hou et al. (2019)
}
