% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_sort_quarterly.R
\name{univariate_sort_quarterly}
\alias{univariate_sort_quarterly}
\title{Perform univariate sorts and compute returns for each portfolio.}
\usage{
univariate_sort_quarterly(
  data,
  var,
  ret = QTRRET,
  n_portfolios = 10,
  nyse_breakpoints = FALSE,
  price_filter = FALSE,
  min_price = 5
)
}
\arguments{
\item{data}{A data frame with quarterly excess returns for each KYPERMNO and the sorting variable \code{var}.}

\item{var}{A variable to sort by.}

\item{ret}{The (excess) return variable, default is MRETRF as MRET - RF.}

\item{n_portfolios}{The number of portfolios to form, default is 10.}

\item{nyse_breakpoints}{Boolean variable for NYSE-based breakpoints, default = \code{FALSE}.}

\item{price_filter}{Boolean variable for applying a screening based on the formation price, default = \code{FALSE}.}

\item{min_price}{\code{dbl} value for the minimum price used in the price screening, default is $5.}
}
\value{
\preformatted{             A tibble with portfolio returns, both equal- and value-weighted on a monthly basis.
}
}
\description{
Sort stocks into \code{n_portfolios} portfolios and compute value-weighted and equal-weighted monthly excess returns.
}
\examples{
\dontrun{
size_deciles <- univariate_sort_quarterly(sfz_mon, QTRTCAP, QTRRETRF)
}
}
